requirements.txt

streamlit
torch
torchvision
Pillow
numpy==1.24.4
